+++ engine/src/pawn_structure.rs
@@ -66,22 +66,22 @@ impl PawnEvaluator {
         // Loop through white pawns
         for sq in white_pawns.iter() {
             // Add the passed pawn mask for this square
-            white_passed_pawns_mask |= self.passed_pawn_mask(Side::White, sq as u8);
+            white_passed_pawns_mask |= self.passed_pawn_mask(Side::White, sq);
 
-            let (file, _rank) = square::from_square(sq as u8);
+            let (file, _rank) = square::from_square(sq);
             if (self.adjacent_file_masks[file as usize] & white_pawns).as_number() == 0 {
-                white_isolated_pawns |= Bitboard::from_square(sq as u8);
+                white_isolated_pawns |= Bitboard::from_square(sq);
             }
         }
 
         // Loop through the white pawns
         for sq in black_pawns.iter() {
             // Add the passed pawn mask for this square
-            black_passed_pawns_mask |= self.passed_pawn_mask(Side::Black, sq as u8);
+            black_passed_pawns_mask |= self.passed_pawn_mask(Side::Black, sq);
 
-            let (file, _rank) = square::from_square(sq as u8);
+            let (file, _rank) = square::from_square(sq);
             if (self.adjacent_file_masks[file as usize] & black_pawns).as_number() == 0 {
-                black_isolated_pawns |= Bitboard::from_square(sq as u8);
+                black_isolated_pawns |= Bitboard::from_square(sq);
             }
         }
